Move unary_op_* functions from `shims::x86::sse` module to `shims::x86`
